+++ modules/parac-modules/library.h
@@ -1,0 +1,23 @@
+//
+// Main header for the Para-C Base Library
+//
+// Created by Luna-K on 08/10/2021.
+//
+
+// local includes
+#include "io.h"
+#include "types.h"
+
+#ifndef PARAC_MODULES_LIBRARY_H
+#define PARAC_MODULES_LIBRARY_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif //PARAC_MODULES_LIBRARY_H